requirements.txt
faiss-cpu==1.8.0.post1
langchain==0.2.12
langchain-community==0.2.10
langchain-google-genai==1.0.7
langgraph==0.2.12
pypdf==4.3.1
docx2txt==0.8
python-dotenv==1.0.1
uvicorn==0.30.3
orjson==3.10.6
Flask==3.0.3

